refactoring: move classes to common module before generalizing them
