+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'checkstyle'
     id 'pmd'
     id 'base'
-    id 'nebula.lint' version '16.22.0'
+    id 'nebula.lint' version '16.23.0'
     id "maven"
     id "io.freefair.lombok" version "5.3.0" apply false
 }
