+++ cristin-import/src/test/java/no/unit/nva/cristin/lambda/CristinEntryEventConsumerTest.java
@@ -19,10 +19,11 @@ import java.io.ByteArrayOutputStream;
 import java.nio.file.Path;
 import java.time.Clock;
 import java.util.Optional;
-import no.unit.nva.cristin.lambda.dtos.FileContentsEvent;
 import no.unit.nva.cristin.mapper.CristinObject;
 import no.unit.nva.events.models.AwsEventBridgeEvent;
 import no.unit.nva.model.Publication;
+import no.unit.nva.publication.s3imports.CristinEntriesEventEmitter;
+import no.unit.nva.publication.s3imports.FileContentsEvent;
 import no.unit.nva.publication.service.ResourcesDynamoDbLocalTest;
 import no.unit.nva.publication.service.impl.ResourceService;
 import no.unit.nva.publication.storage.model.UserInstance;
