+++ s3-import-commons/build.gradle
@@ -1,13 +1,10 @@
-apply plugin: 'io.freefair.lombok'
-
 dependencies {
     implementation project(":publication-constants")
     implementation project(":publication-commons")
-    implementation project(":s3-import-commons")
 
-    implementation(group: 'com.github.bibsysdev', name: 'nva-datamodel-java', version: project.ext.nvaDatamodelVersion) {
-        exclude group: 'com.github.bibsysdev', module: 'nva-commons'
-    }
+//    implementation(group: 'com.github.bibsysdev', name: 'nva-datamodel-java', version: project.ext.nvaDatamodelVersion) {
+//        exclude group: 'com.github.bibsysdev', module: 'nva-commons'
+//    }
 
     implementation group: 'com.github.bibsysdev', name: 's3', version: project.ext.nvaCommonsVersion
     implementation group: 'com.github.bibsysdev', name: 'core', version: project.ext.nvaCommonsVersion
@@ -16,15 +13,15 @@ dependencies {
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: project.ext.jacksonVersion
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: project.ext.jacksonVersion
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: project.ext.jacksonVersion
+
+    implementation group: 'software.amazon.awssdk', name: 'aws-core', version: project.ext.awsSdk2Version
+    implementation group: 'software.amazon.awssdk', name: 'sdk-core', version: project.ext.awsSdk2Version
+
+    implementation group: 'software.amazon.awssdk', name: 'http-client-spi', version: project.ext.awsSdk2Version
+    implementation group: 'org.apache.commons', name: 'commons-collections4', version: '4.4'
+
     runtimeOnly group: "com.fasterxml.jackson.datatype", name: "jackson-datatype-jsr310", version: project.ext.jacksonVersion
     runtimeOnly group: "com.fasterxml.jackson.datatype", name: "jackson-datatype-jdk8", version: project.ext.jacksonVersion
-
-    implementation group: 'com.amazonaws', name: 'aws-java-sdk-dynamodb', version: project.ext.awsSdkVersion
-    implementation group: 'com.amazonaws', name: 'aws-java-sdk-core', version: project.ext.awsSdkVersion
-
-    implementation group: 'software.amazon.awssdk', name: 'regions', version: project.ext.awsSdk2Version
-
-    runtimeOnly group: 'software.amazon.awssdk', name: 'url-connection-client', version: project.ext.awsSdk2Version
 
     testImplementation project(":storage-model")
     testImplementation project(":publication-testing")
@@ -32,31 +29,32 @@ dependencies {
     testImplementation group: 'org.mockito', name: 'mockito-core', version: project.ext.mockitoVersion
     testImplementation group: 'com.github.bibsysdev', name: 'nva-testutils', version: project.ext.nvaTestUtilsVersion
     testImplementation group: 'com.github.bibsysdev', name: 'logutils', version: project.ext.nvaCommonsVersion
-    testImplementation group: 'com.github.bibsysdev', name: 'identifiers', version: project.ext.nvaCommonsVersion
-    testImplementation(group: 'org.javers', name: 'javers-core', version: project.ext.javersVersion) {
-        because "For comparing java objects and detailing the differences in readable way"
-    }
+    testImplementation group: 'software.amazon.awssdk', name: 'utils', version: project.ext.awsSdk2Version
+    testImplementation group: 'software.amazon.awssdk', name: 'eventbridge', version: '2.16.46'
     testImplementation group: 'software.amazon.awssdk', name: 's3', version: project.ext.awsSdk2Version
-
-    compileOnly group: 'software.amazon.awssdk', name: 'eventbridge', version: '2.16.46'
-    compileOnly group: 'software.amazon.awssdk', name: 's3', version: project.ext.awsSdk2Version
 
     compileOnly(project(":storage-model")) {
         because "PMD complains for no/unit/nva/publication/storage/model/daos/DynamoEntry"
     }
     compileOnly group: 'com.github.bibsysdev', name: 'apigateway', version: project.ext.nvaCommonsVersion
-    compileOnly group: 'org.zalando', name: 'problem', version: project.ext.zalandoProblemVersion
-    compileOnly group: 'com.amazonaws', name: 'aws-lambda-java-core', version: project.ext.awsLambdaCore
     compileOnly group: 'com.github.bibsysdev', name: 'identifiers', version: project.ext.nvaCommonsVersion
+    compileOnly group: 'org.zalando', name: 'problem', version: project.ext.zalandoProblemVersion
+
+    compileOnly group: 'com.amazonaws', name: 'aws-lambda-java-core', version: project.ext.awsLambdaCore
+    compileOnly group: 'software.amazon.awssdk', name: 'eventbridge', version: '2.16.46'
+    compileOnly group: 'software.amazon.awssdk', name: 's3', version: project.ext.awsSdk2Version
+    compileOnly group: 'software.amazon.awssdk', name: 'url-connection-client', version: project.ext.awsSdk2Version
+    compileOnly group: 'software.amazon.awssdk', name: 'regions', version: project.ext.awsSdk2Version
 
     compileOnly(group: 'software.amazon.awssdk', name: 'auth', version: project.ext.awsSdk2Version) {
         because "PMD complains about software/amazon/awssdk/auth/credentials/AwsCredentialsProvider "
     }
-
+    compileOnly(group: 'com.amazonaws', name: 'aws-java-sdk-api-gateway', version: project.ext.awsSdkVersion) {
+        because "PMD complains for the com/amazonaws/regions/Regions class"
+    }
     compileOnly(group: 'commons-validator', name: 'commons-validator', version: '1.7') {
         because "PMD org/apache/commons/validator/routines/ISBNValidator"
     }
-
 }
 
 test {
