+++ s3-import-commons/src/test/java/no/unit/nva/publication/s3imports/FakeEventBridgeClient.java
@@ -1,10 +1,9 @@
-package no.unit.nva.cristin.lambda;
+package no.unit.nva.publication.s3imports;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import no.unit.nva.cristin.lambda.dtos.ImportRequest;
 import software.amazon.awssdk.awscore.exception.AwsServiceException;
 import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.services.eventbridge.EventBridgeClient;
