+++ s3-import-commons/src/test/java/no/unit/nva/publication/s3imports/SampleObject.java
@@ -1,4 +1,4 @@
-package no.unit.nva.cristin.lambda;
+package no.unit.nva.publication.s3imports;
 
 import static nva.commons.core.attempt.Try.attempt;
 import com.fasterxml.jackson.annotation.JsonCreator;
