+++ settings.gradle
@@ -6,6 +6,7 @@ include 'publication-commons'
 include 'doi-requests'
 include 'messages'
 include 'create-publication'
+include 's3-import-commons'
 include 'data-import-from-s3'
 include 'data-migration'
 include 'publication-events'
